% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_models_patch_level.R
\name{all_models_patch_level}
\alias{all_models_patch_level}
\title{Get results for all 4 models (ODE and SDE) at patch level}
\usage{
all_models_patch_level(
  carrying_capacity,
  n0 = carrying_capacity,
  t_max = 5,
  repetitions = 250
)
}
\arguments{
\item{carrying_capacity}{vector of carrying capacity}

\item{n0}{vector of starting numbers}
}
\description{
Get results for all 4 models (ODE and SDE) at patch level
}
