% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveLandscape-package.R
\docType{package}
\name{curveLandscape-package}
\alias{curveLandscape}
\alias{curveLandscape-package}
\title{curveLandscape: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mossa Merhi Reimert \email{mossa@sund.ku.dk}

}
\keyword{internal}
