% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_linear_id}
\alias{get_linear_id}
\title{Returns the linear, 0-index id for (i,j) for n x n matrix.}
\usage{
get_linear_id(i, j, n)
}
\description{
Returns the linear, 0-index id for (i,j) for n x n matrix.
}
