% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_row_col}
\alias{get_row_col}
\title{Returns the row and column from a 0-indexed, column-wise linear index \code{k} for a square matrix of dimension \code{n} x \code{n}}
\usage{
get_row_col(k, n)
}
\description{
Returns the row and column from a 0-indexed, column-wise linear index \code{k} for a square matrix of dimension \code{n} x \code{n}
}
