% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_total_number_of_elements}
\alias{get_total_number_of_elements}
\title{Return the total number of elements in lower-triangular matrix (without diagonal)}
\usage{
get_total_number_of_elements(n)
}
\description{
Return the total number of elements in lower-triangular matrix (without diagonal)
}
