% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sim_bd_only}
\alias{sim_bd_only}
\title{Simulates a birth and death process that is density-dependent}
\usage{
sim_bd_only(n0, birth_baseline, death_baseline, carrying_capacity, t_max)
}
\description{
Simulates a birth and death process that is density-dependent
}
