% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sim_migration_only}
\alias{sim_migration_only}
\title{Perform only simulation with migration, no birth/death feedback}
\usage{
sim_migration_only(n0, migration_baseline, carrying_capacity, k_dij, t_max)
}
\description{
Perform only simulation with migration, no birth/death feedback
}
