% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{update_migration_only}
\alias{update_migration_only}
\title{Updates the population state vector according to the migration equation}
\usage{
update_migration_only(
  population_total,
  migration_baseline,
  carrying_capacity,
  k_dij
)
}
\description{
(CURRENT) Equation (1 - exp) m0 • k(d_ij) • exp(-cc\link{source})
(MISSING) Equation (1 - div) m0 • k(d_ij) / cc\link{source}
(MISSING) Equation (2 - exp) m0 • k(d_ij) • exp(-cc\link{source}) • exp(-(cc\link{target} - n\link{source}))
(MISSING) Equation (2 - div) m0 • k(d_ij) n\link{source} / (cc\link{source}) • cc\link{target})
}
