% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sim_bdm}
\alias{sim_bdm}
\title{Simulates birth, death and migration process of a multi-patch system.}
\usage{
sim_bdm(n0, birth_baseline, death_baseline, t_max)
}
\description{
Simulates birth, death and migration process of a multi-patch system.
}
